requirements.txt
aiohttp
beautifulsoup4
pytest-playwright